Extract pack for support
