+++ src/test/java/atdd/station/EdgeAcceptanceTest.java
@@ -1,5 +1,8 @@
 package atdd.station;
 
+import atdd.station.support.EdgeHttpSupport;
+import atdd.station.support.LineHttpSupport;
+import atdd.station.support.StationHttpSupport;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
