+++ src/test/java/atdd/station/LineAcceptanceTest.java
@@ -1,6 +1,7 @@
 package atdd.station;
 
 import atdd.line.LineDto;
+import atdd.station.support.LineHttpSupport;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
