+++ src/test/java/atdd/station/StationAcceptanceTest.java
@@ -1,5 +1,6 @@
 package atdd.station;
 
+import atdd.station.support.StationHttpSupport;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
