+++ src/test/java/atdd/station/support/EdgeHttpSupport.java
@@ -1,28 +1,33 @@
-package atdd.station;
+package atdd.station.support;
 
-import atdd.line.Line;
-import atdd.line.LineLink;
+import atdd.edge.Edge;
+import atdd.edge.EdgeLink;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.reactive.server.EntityExchangeResult;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import reactor.core.publisher.Mono;
 
-public class LineHttpSupport {
+public class EdgeHttpSupport {
 
+    public static EntityExchangeResult<Edge> createEdge(WebTestClient webTestClient) {
 
-    public static EntityExchangeResult<Line> create(WebTestClient webTestClient) {
+        final String inputJson = "{\n" +
+                "  \"lineId\": 1,\n" +
+                "  \"elapsedTime\": 5,\n" +
+                "  \"distance\": 2.0,\n" +
+                "  \"sourceStationId\": 2,\n" +
+                "  \"targetStationId\": 3\n" +
+                "}";
 
-        final String inputJson = "{\"name\":\"1호선\",\"startTime\":\"05:30\",\"endTime\":\"23:30\",\"intervalTime\":5}";
-
-        return webTestClient.post().uri(LineLink.ROOT)
+        return webTestClient.post().uri(EdgeLink.ROOT)
                 .contentType(MediaType.APPLICATION_JSON)
                 .body(Mono.just(inputJson), String.class)
                 .exchange()
                 .expectStatus().isCreated()
                 .expectHeader().contentType(MediaType.APPLICATION_JSON)
                 .expectHeader().exists(HttpHeaders.LOCATION)
-                .expectBody(Line.class)
+                .expectBody(Edge.class)
                 .returnResult();
     }
 }
