+++ src/test/java/atdd/station/support/LineHttpSupport.java
@@ -1,33 +1,28 @@
-package atdd.station;
+package atdd.station.support;
 
-import atdd.edge.Edge;
-import atdd.edge.EdgeLink;
+import atdd.line.Line;
+import atdd.line.LineLink;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.reactive.server.EntityExchangeResult;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import reactor.core.publisher.Mono;
 
-public class EdgeHttpSupport {
+public class LineHttpSupport {
 
-    public static EntityExchangeResult<Edge> createEdge(WebTestClient webTestClient) {
 
-        final String inputJson = "{\n" +
-                "  \"lineId\": 1,\n" +
-                "  \"elapsedTime\": 5,\n" +
-                "  \"distance\": 2.0,\n" +
-                "  \"sourceStationId\": 2,\n" +
-                "  \"targetStationId\": 3\n" +
-                "}";
+    public static EntityExchangeResult<Line> create(WebTestClient webTestClient) {
 
-        return webTestClient.post().uri(EdgeLink.ROOT)
+        final String inputJson = "{\"name\":\"1호선\",\"startTime\":\"05:30\",\"endTime\":\"23:30\",\"intervalTime\":5}";
+
+        return webTestClient.post().uri(LineLink.ROOT)
                 .contentType(MediaType.APPLICATION_JSON)
                 .body(Mono.just(inputJson), String.class)
                 .exchange()
                 .expectStatus().isCreated()
                 .expectHeader().contentType(MediaType.APPLICATION_JSON)
                 .expectHeader().exists(HttpHeaders.LOCATION)
-                .expectBody(Edge.class)
+                .expectBody(Line.class)
                 .returnResult();
     }
 }
